refactor(agent): extract RagOrchestrator + characterization test (PR-2a slice 5)
