+++ packages/llm-agent-libs/src/agent.ts
@@ -45,6 +45,7 @@ import {
   toToolCallDelta,
 } from '@mcp-abap-adt/llm-agent';
 import { wrapEmbedder } from './adapters/usage-logging-embedder.js';
+import { RagOrchestrator } from './agent/rag-orchestrator.js';
 import { normalizeRequestOptions } from './agent-request-options.js';
 import type { LlmClassifierConfig } from './classifier/llm-classifier.js';
 import { LlmClassifier } from './classifier/llm-classifier.js';
@@ -664,348 +665,57 @@ export class SmartAgent {
         return;
       }
 
-      // 1. Unified Preparation (default hardcoded flow)
-      const initResult = await this._preparePipeline(
-        textOrMessages,
+      // Default hardcoded flow: RAG fan-out + context assembly. The orchestrator
+      // is constructed PER REQUEST so it reads the LIVE _mainLlm/_helperLlm/
+      // _classifier (hot-swap via reconfigure() keeps working — see issue #164).
+      const orchestrator = new RagOrchestrator({
+        mainLlm: this._mainLlm,
+        helperLlm: this._helperLlm,
+        classifier: this._classifier,
+        config: this.config,
+        tracer: this.tracer,
+        metrics: this.metrics,
+        reranker: this.reranker,
+        queryExpander: this.queryExpander,
+        sessionManager: this.sessionManager,
+        toolAvailabilityRegistry: this.toolAvailabilityRegistry,
+        mcpToolRegistry: this.mcpToolRegistry,
+        requestLogger: this.requestLogger,
+        ragStores: this.deps.ragStores,
+        embedder: this.deps.embedder,
+        assembler: this.deps.assembler,
+        skillManager: this.deps.skillManager,
+        translateQueryStores: this.deps.translateQueryStores,
+      });
+      const orchResult = await orchestrator.orchestrate(textOrMessages, {
         opts,
         rootSpan,
-      );
-      if (!initResult.ok) {
-        rootSpan.setStatus('error', initResult.error.message);
-        rootSpan.end();
-        yield initResult;
-        return;
-      }
-      let { processedHistory } = initResult.value;
-      const { subprompts, toolClientMap } = initResult.value;
-
-      // Token budget check — summarize if over budget
-      if (this.sessionManager.isOverBudget()) {
-        const sumResult = await this._summarizeHistory(processedHistory, opts);
-        if (sumResult.ok) processedHistory = sumResult.value;
-        this.sessionManager.reset();
-      }
-
-      // 2. Decide context and tools for the WHOLE request
-      await this.mcpToolRegistry.resolveActiveClients(opts);
-      const actions = subprompts.filter((sp) => sp.type === 'action');
-      const hasActions = actions.length > 0;
-      const hasMcpClients = this.mcpToolRegistry.getActiveClients().length > 0;
-      const hasRagStores = Object.keys(this.deps.ragStores).length > 0;
-      const shouldRetrieve =
-        mode === 'hard' || (hasActions && (hasMcpClients || hasRagStores));
-
-      let finalTools: LlmTool[] = [];
-      let retrieved: {
-        ragResults: Record<string, RagResult[]>;
-        tools: McpTool[];
-      } = {
-        ragResults: {},
-        tools: [],
-      };
-      let skillContent = '';
-
-      if (shouldRetrieve) {
-        // Collect all action texts for RAG
-        const combinedActionText = actions.map((a) => a.text).join(' ');
-
-        // Translate + expand once (only used for stores in translateQueryStores)
-        const translateStores = this.deps.translateQueryStores;
-        let translatedText: string | undefined;
-        if (translateStores && translateStores.size > 0) {
-          translatedText = await this._toEnglishForRag(
-            combinedActionText,
-            opts,
-          );
-          if (this.config.queryExpansionEnabled) {
-            const expandResult = await this.queryExpander.expand(
-              translatedText,
-              opts,
-            );
-            if (expandResult.ok) translatedText = expandResult.value;
-          }
-        }
-
-        const k = this.config.ragQueryK ?? 10;
-        const ragSpan = this.tracer.startSpan('smart_agent.rag_query', {
-          parent: rootSpan,
-          attributes: { 'rag.k': k },
-        });
-        const storeEntries = Object.entries(this.deps.ragStores);
-
-        // Build per-store embedding: translated for translateQuery stores, original for others
-        const mkEmbed = (text: string) =>
-          this.deps.embedder
-            ? new QueryEmbedding(text, this.deps.embedder, opts)
-            : new TextOnlyEmbedding(text);
-        // Cache embeddings to avoid duplicate embed calls
-        const originalEmbedding = mkEmbed(combinedActionText);
-        const translatedEmbedding =
-          translatedText && translatedText !== combinedActionText
-            ? mkEmbed(translatedText)
-            : originalEmbedding;
-
-        const ragQueryResults = await Promise.all(
-          storeEntries.map(([name, store]) => {
-            const emb =
-              translateStores?.has(name) && translatedText
-                ? translatedEmbedding
-                : originalEmbedding;
-            return store.query(emb, k, opts).then((r) => ({ name, result: r }));
-          }),
-        );
-        ragSpan.end();
-        const ragResultsMap: Record<string, RagResult[]> = {};
-        for (const { name, result: r } of ragQueryResults) {
-          ragResultsMap[name] = r.ok ? r.value : [];
-          this.metrics.ragQueryCount.add(1, {
-            store: name,
-            hit: String(r.ok && r.value.length > 0),
-          });
-        }
-
-        // Rerank results
-        // Rerank all stores in parallel, using matching query text per store
-        const rerankedEntries = await Promise.all(
-          Object.entries(ragResultsMap).map(async ([name, results]) => {
-            if (results.length > 0) {
-              const rerankText =
-                translateStores?.has(name) && translatedText
-                  ? translatedText
-                  : combinedActionText;
-              const rr = await this.reranker.rerank(rerankText, results, opts);
-              return { name, results: rr.ok ? rr.value : results };
-            }
-            return { name, results };
-          }),
-        );
-        const rerankedMap: Record<string, RagResult[]> = {};
-        for (const { name, results } of rerankedEntries) {
-          rerankedMap[name] = results;
-        }
-
-        const { tools: mcpTools } = await this.mcpToolRegistry.resolve(opts);
-
-        // Collect all RAG results for tool discovery
-        const allRagResults = Object.values(rerankedMap).flat();
-
-        // Log RAG results with scores for diagnostics
-        for (const [storeName, results] of Object.entries(rerankedMap)) {
-          const logQuery =
-            translateStores?.has(storeName) && translatedText
-              ? translatedText
-              : combinedActionText;
-          opts?.sessionLogger?.logStep(`rag_query_${storeName}`, {
-            query: logQuery.slice(0, 200),
-            k,
-            resultCount: results.length,
-            results: results.map((r) => ({
-              id: r.metadata.id,
-              score: r.score,
-              text: r.text.slice(0, 120),
-            })),
-          });
-        }
-
-        const ragToolNames = new Set(
-          allRagResults
-            .map((r) => r.metadata.id as string)
-            .filter((id) => id?.startsWith('tool:'))
-            .map((id) => id.slice(5).replace(/:.*$/, '')),
-        );
-        const selectedMcpTools =
-          ragToolNames.size > 0
-            ? mcpTools.filter((t) => ragToolNames.has(t.name))
-            : mode === 'hard'
-              ? mcpTools
-              : [];
-
-        // Log tool selection diagnostics
-        opts?.sessionLogger?.logStep('tools_selected', {
-          totalMcp: mcpTools.length,
-          ragMatchedTools: [...ragToolNames],
-          selectedCount: selectedMcpTools.length + externalTools.length,
-          selectedNames: [
-            ...selectedMcpTools.map((t) => t.name),
-            ...externalTools.map((t) => t.name),
-          ],
-        });
-
-        retrieved = {
-          ragResults: rerankedMap,
-          tools: selectedMcpTools,
-        };
-        // D4: external (client) tools are always offered regardless of mode;
-        // mode governs only the worker's INTERNAL execution posture.
-        finalTools = [...(selectedMcpTools as LlmTool[]), ...externalTools];
-        opts?.sessionLogger?.logStep('external_tools_merge', {
-          mode,
-          mcpCount: selectedMcpTools.length,
-          externalCount: externalTools.length,
-          externalNames: externalTools.map((t) => t.name),
-          finalCount: finalTools.length,
-        });
-
-        // Skill injection (when enabled and skillManager configured)
-        if (
-          this.config.skillInjectionEnabled !== false &&
-          this.deps.skillManager
-        ) {
-          const ragSkillNames = new Set(
-            allRagResults
-              .map((r) => r.metadata.id as string)
-              .filter((id) => id?.startsWith('skill:'))
-              .map((id) => id.slice(6)),
-          );
-
-          // Fallback: dedicated RAG query when no skill:* in existing results
-          if (ragSkillNames.size === 0) {
-            const k = this.config.ragQueryK ?? 15;
-            const storeEntries = Object.entries(this.deps.ragStores);
-            const fallbackResults = await Promise.all(
-              storeEntries.map(([name, store]) => {
-                const text =
-                  translateStores?.has(name) && translatedText
-                    ? translatedText
-                    : combinedActionText;
-                const emb = this.deps.embedder
-                  ? new QueryEmbedding(text, this.deps.embedder, opts)
-                  : new TextOnlyEmbedding(text);
-                return store.query(emb, k, opts);
-              }),
-            );
-            for (const result of fallbackResults) {
-              if (result.ok) {
-                for (const r of result.value) {
-                  const id = r.metadata.id as string;
-                  if (id?.startsWith('skill:')) {
-                    ragSkillNames.add(id.slice(6));
-                  }
-                }
-              }
-            }
-            if (ragSkillNames.size > 0) {
-              opts?.sessionLogger?.logStep('skill_select_rag_fallback', {
-                query: combinedActionText.slice(0, 200),
-                k,
-                matchedSkills: [...ragSkillNames],
-              });
-            }
-          }
-
-          const allSkillsResult = await this.deps.skillManager.listSkills(opts);
-          if (allSkillsResult.ok) {
-            const allSkills = allSkillsResult.value;
-            const matched =
-              ragSkillNames.size > 0
-                ? allSkills.filter((s) => ragSkillNames.has(s.name))
-                : mode === 'hard'
-                  ? allSkills
-                  : [];
-            const contentParts: string[] = [];
-            for (const skill of matched) {
-              const contentResult = await skill.getContent(undefined, opts);
-              if (contentResult.ok && contentResult.value) {
-                contentParts.push(
-                  `### Skill: ${skill.name}\n${contentResult.value}`,
-                );
-              }
-            }
-            skillContent = contentParts.join('\n\n');
-            opts?.sessionLogger?.logStep('skills_selected', {
-              totalSkills: allSkills.length,
-              ragMatchedSkills: [...ragSkillNames],
-              selectedCount: matched.length,
-              selectedNames: matched.map((s) => s.name),
-            });
-          }
-        }
-      } else {
-        // If we're here, mode is definitely 'smart' (not 'hard' or 'pass')
-        finalTools = externalTools;
-      }
-      const filteredTools = this.toolAvailabilityRegistry.filterTools(
         sessionId,
-        finalTools,
-      );
-      finalTools = filteredTools.allowed;
-      if (filteredTools.blocked.length > 0) {
-        opts?.sessionLogger?.logStep('active_tools_filtered_by_registry', {
-          blocked: filteredTools.blocked,
-        });
-      }
-
-      // 3. Assemble Context once
-      const mainAction =
-        actions.length > 1
-          ? {
-              type: 'action' as const,
-              text: actions.map((a) => a.text).join('\n'),
-              context: actions.find((a) => a.context)?.context,
-              dependency: 'independent' as const,
-            }
-          : actions.length === 1
-            ? actions[0]
-            : subprompts.find((sp) => sp.type === 'chat') || subprompts[0];
-
-      if (actions.length > 1) {
-        opts?.sessionLogger?.logStep('actions_merged', {
-          count: actions.length,
-          actions: actions.map((a) => ({
-            text: a.text,
-            dependency: a.dependency,
-          })),
-        });
-      }
-      const assembleSpan = this.tracer.startSpan('smart_agent.assemble', {
-        parent: rootSpan,
+        mode,
+        externalTools,
       });
-      const assembleResult = await this.deps.assembler.assemble(
-        mainAction,
-        retrieved,
-        processedHistory,
-        opts,
-      );
-      if (!assembleResult.ok) {
-        assembleSpan.setStatus('error', assembleResult.error.message);
-        assembleSpan.end();
-        rootSpan.setStatus('error', assembleResult.error.message);
+      if (!orchResult.ok) {
+        rootSpan.setStatus('error', orchResult.error.message);
         rootSpan.end();
-        yield {
-          ok: false,
-          error: new OrchestratorError(
-            assembleResult.error.message,
-            'ASSEMBLER_ERROR',
-          ),
-        };
+        yield orchResult;
         return;
       }
-      assembleSpan.setStatus('ok');
-      assembleSpan.end();
-
-      // Inject skill content into system message (post-assembly)
-      if (skillContent) {
-        const sysMsg = assembleResult.value.find((m) => m.role === 'system');
-        if (sysMsg) {
-          sysMsg.content += `\n\n## Active Skills\n${skillContent}`;
-        } else {
-          assembleResult.value.unshift({
-            role: 'system' as const,
-            content: `## Active Skills\n${skillContent}`,
-          });
-        }
-      }
-
-      opts?.sessionLogger?.logStep(`final_context_assembled`, {
-        messages: assembleResult.value,
-        tools: finalTools.map((t) => t.name),
-      });
+      // skillContent is NOT destructured — the caller doesn't use it; it is
+      // already baked into assembledMessages inside orchestrate(). Destructuring
+      // it unused would trip noUnusedLocals.
+      const {
+        retrieved,
+        finalTools,
+        assembledMessages,
+        mainAction,
+        toolClientMap,
+      } = orchResult.value;
 
       // 4. Single Streaming Loop
       const stream = this._runStreamingToolLoop(
         mainAction,
         retrieved,
-        assembleResult.value,
+        assembledMessages,
         toolClientMap,
         opts,
         rootSpan,
@@ -1022,72 +732,6 @@ export class SmartAgent {
       timeoutCleanup?.();
       this.metrics.requestLatency.record(Date.now() - requestStart);
     }
-  }
-
-  private async _preparePipeline(
-    textOrMessages: string | Message[],
-    opts: CallOptions | undefined,
-    parentSpan: ISpan,
-  ): Promise<
-    Result<
-      {
-        subprompts: Subprompt[];
-        processedHistory: Message[];
-        toolClientMap: Map<string, IMcpClient>;
-      },
-      OrchestratorError
-    >
-  > {
-    opts?.sessionLogger?.logStep('client_request', { textOrMessages });
-    const text =
-      typeof textOrMessages === 'string'
-        ? textOrMessages
-        : (textOrMessages.filter((m) => m.role === 'user').slice(-1)[0]
-            ?.content ?? '');
-    const history = typeof textOrMessages === 'string' ? [] : textOrMessages;
-    let processedHistory = history;
-    const summarizeLimit = this.config.historyAutoSummarizeLimit ?? 10;
-    if (this._helperLlm && history.length > summarizeLimit) {
-      const res = await this._summarizeHistory(history, opts);
-      if (res.ok) processedHistory = res.value;
-    }
-
-    let subprompts: Subprompt[];
-
-    if (this.config.classificationEnabled === false) {
-      // Skip classification — treat entire input as a single action
-      subprompts = [
-        { type: 'action', text, dependency: 'independent' as const },
-      ];
-      opts?.sessionLogger?.logStep('classification_skipped', { text });
-    } else {
-      const classifySpan = this.tracer.startSpan('smart_agent.classify', {
-        parent: parentSpan,
-      });
-      const classifyResult = await this._classifier.classify(text, opts);
-      if (!classifyResult.ok) {
-        classifySpan.setStatus('error', classifyResult.error.message);
-        classifySpan.end();
-        return {
-          ok: false,
-          error: new OrchestratorError(
-            classifyResult.error.message,
-            'CLASSIFIER_ERROR',
-          ),
-        };
-      }
-      classifySpan.setStatus('ok');
-      classifySpan.end();
-      opts?.sessionLogger?.logStep('classifier_response', {
-        subprompts: classifyResult.value,
-      });
-      subprompts = classifyResult.value;
-    }
-    for (const sp of subprompts) {
-      this.metrics.classifierIntentCount.add(1, { intent: sp.type });
-    }
-    const { toolClientMap } = await this.mcpToolRegistry.resolve(opts);
-    return { ok: true, value: { subprompts, processedHistory, toolClientMap } };
   }
 
   private async *_runStreamingToolLoop(
@@ -1852,71 +1496,5 @@ export class SmartAgent {
       }
       messages = [...messages, ...toolMessages];
     }
-  }
-
-  private async _toEnglishForRag(
-    text: string,
-    opts: CallOptions | undefined,
-  ): Promise<string> {
-    if (/^[\p{ASCII}]+$/u.test(text) || text.length < 15) return text;
-    const dp =
-      'Translate the user request to English for search purposes. Preserve technical terms if present. Reply with only the expanded English terms, no explanation.';
-    const llm = this._helperLlm || this._mainLlm;
-    const res = await llm.chat(
-      [
-        {
-          role: 'system' as const,
-          content: this.config.ragTranslatePrompt || dp,
-        },
-        { role: 'user' as const, content: text },
-      ],
-      [],
-      opts,
-    );
-    return res.ok && res.value.content.trim() ? res.value.content.trim() : text;
-  }
-
-  private async _summarizeHistory(
-    h: Message[],
-    opts?: CallOptions,
-  ): Promise<Result<Message[], OrchestratorError>> {
-    if (!this._helperLlm) return { ok: true, value: h };
-    const toS = h.slice(0, -5);
-    const rec = h.slice(-5);
-    if (toS.length === 0) return { ok: true, value: h };
-    const dp =
-      'Summarize the conversation so far in 2-3 sentences. Focus on the user goals and the current status of the task. Keep technical SAP terms as is.';
-    const summarizeStart = Date.now();
-    const res = await this._helperLlm.chat(
-      [
-        ...toS,
-        {
-          role: 'system' as const,
-          content: this.config.historySummaryPrompt || dp,
-        },
-      ],
-      [],
-      opts,
-    );
-    this.requestLogger.logLlmCall({
-      component: 'helper',
-      model: this._helperLlm.model ?? 'unknown',
-      promptTokens: res.ok ? (res.value.usage?.promptTokens ?? 0) : 0,
-      completionTokens: res.ok ? (res.value.usage?.completionTokens ?? 0) : 0,
-      totalTokens: res.ok ? (res.value.usage?.totalTokens ?? 0) : 0,
-      durationMs: Date.now() - summarizeStart,
-      requestId: opts?.trace?.traceId,
-    });
-    if (!res.ok) return { ok: true, value: h };
-    return {
-      ok: true,
-      value: [
-        {
-          role: 'system' as const,
-          content: `Summary of previous conversation: ${res.value.content}`,
-        },
-        ...rec,
-      ],
-    };
   }
 }
